refactor : TeamResponseUtils로 구조 개선
