+++ src/main/java/site/codemonster/comon/domain/article/service/ArticleService.java
@@ -1,10 +1,8 @@
 package site.codemonster.comon.domain.article.service;
 
 
-import java.util.Optional;
 import site.codemonster.comon.domain.article.dto.response.ArticleParticularDateResponse;
 import site.codemonster.comon.domain.article.dto.response.ArticleResponse;
-import site.codemonster.comon.domain.article.dto.response.TeamSubjectResponse;
 import site.codemonster.comon.domain.article.factory.RecommendationArticleFactory;
 import site.codemonster.comon.domain.article.dto.request.*;
 import site.codemonster.comon.domain.article.entity.Article;
@@ -16,6 +14,7 @@ import site.codemonster.comon.domain.problem.entity.Problem;
 import site.codemonster.comon.domain.team.dto.response.TeamPageResponse;
 import site.codemonster.comon.domain.team.entity.Team;
 import site.codemonster.comon.domain.team.service.TeamService;
+import site.codemonster.comon.domain.team.utils.TeamResponseUtils;
 import site.codemonster.comon.domain.teamMember.service.TeamMemberService;
 import site.codemonster.comon.global.error.Team.TeamManagerInvalidException;
 import site.codemonster.comon.global.error.articles.*;
@@ -42,6 +41,7 @@ public class ArticleService {
     private final ArticleRepository articleRepository;
     private final ArticleImageRepository articleImageRepository;
     private final ImageFieldConvertUtils imageFieldConvertUtils;
+    private final TeamResponseUtils teamResponseUtils;
 
     @Transactional
     public Article articleCreate(Member member, ArticleCreateRequest articleCreateRequest) {
@@ -152,7 +152,7 @@ public class ArticleService {
 
         List<Article> subjectArticles = articleRepository.findSubjectArticlesByTeamIdAndYearAndMonth(teamId, calenderSubjectRequest.year(), calenderSubjectRequest.month(), getSubjectCategories());
         boolean isTeamManager = teamMemberService.checkMemberIsTeamManager(teamId, member);
-        return TeamPageResponse.from(teamService.getMyTeamResponse(team), isTeamManager, subjectArticles);
+        return TeamPageResponse.from(teamResponseUtils.getMyTeamResponse(team), isTeamManager, subjectArticles);
     }
 
     private void validateTeamManager(Member member, Team team) {
